Move F1 to its own file
